Update darwin_m1_setup.sh
- Virtual environment setup is now done manually
- Fix protobuf library bug where "libprotobuf.11.dylib is not found"

+++ setup/darwin_m1_setup.sh
@@ -5,14 +5,18 @@
 ### Install Tensorflow
 echo "Installing tensorflow-macos and tensorflow-metal..."
 
-# Creating a virtual environment and install pip
-python3 -m venv ~/tensorflow-metal
-source ~/tensorflow-metal/bin/activate
+# Install pip
 python -m pip install -U pip
 
-# Download and install conda env
-curl https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-MacOSX-arm64.sh -o Miniforge3-MacOSX-arm64.sh
-bash Miniforge3-MacOSX-arm64.sh
+# Download and install conda environment
+if [ -f "Miniforge3-MacOSX-arm64.sh" ]; then
+    echo "Miniforge3-MacOSX-arm64.sh detected."
+    bash Miniforge3-MacOSX-arm64.sh
+else
+    echo "Miniforge3-MacOSX-arm64.sh not detected. Downloading..."
+    curl https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-MacOSX-arm64.sh -o Miniforge3-MacOSX-arm64.sh
+    bash Miniforge3-MacOSX-arm64.sh
+fi
 
 # Uninstall existing tensorflow-macos and tensorflow-metal
 python -m pip uninstall tensorflow-macos
@@ -29,19 +33,10 @@ python -m pip install tensorflow-metal
 
 ### Install tf2onnx
 echo "Installing tf2onnx..."
-
 python -m pip install tf2onnx
 
 ### Install onnx
 echo "Installing onnx..."
-
-# Reinstall cmake and protobuf
-brew uninstall cmake
-brew install cmake
-
-brew uninstall protobuf
-brew install protobuf
-
 python -m pip install onnx
 
 # Sucess message
